modularizacion de barra de navegacion
